+++ addons/yat/src/scenes/monitor/components/memory_info/MemoryInfo.cs
@@ -23,17 +23,19 @@ namespace YAT.Scenes.Monitor
 			var physical = mem["physical"];
 			var free = mem["free"];
 			var stack = mem["stack"];
+			var vram = Performance.GetMonitor(Performance.Monitor.RenderVideoMemUsed);
 
 			var freePercent = (float)free / (float)physical * 100f;
 
 			_label.Clear();
 			_label.AppendText(
-				$"Memory: {NumericHelper.SizeToString(physical.AsInt64(), 3)}\n" +
+				$"RAM: {NumericHelper.SizeToString(physical.AsInt64(), 3)}\n" +
 				$"Free: {(
-					freePercent < 15
+					freePercent < 15 && UseColors
 					? $"[color={_yat.Options.ErrorColor}]{NumericHelper.SizeToString(free.AsInt64(), 3)}[/color]"
 					: NumericHelper.SizeToString(free.AsInt64(), 3))}\n" +
-				$"Stack: {NumericHelper.SizeToString(stack.AsInt64(), 1)}"
+				$"Stack: {NumericHelper.SizeToString(stack.AsInt64(), 1)}\n" +
+				$"VRAM: {NumericHelper.SizeToString((long)vram, 2)}"
 			);
 		}
 	}